+++ src/main/java/Application.java
@@ -1,3 +1,5 @@
+import controller.Controller;
+
 public class Application {
     public static void main(String[] args) {
         Controller controller = new Controller();
