+++ src/main/java/domain/game/DealerResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 import java.util.Map;
 import java.util.stream.Collectors;
