+++ src/main/java/domain/game/State.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.game;
 
 public enum State {
     HIT, STAY, BUST;
